Responsive design update
