+++ lib/controllers/home_controller.dart
@@ -1,18 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
-import 'package:speedwatch/components/rightpane.dart';
-import 'package:speedwatch/components/sessions_list.dart';
-
-import '../components/sidebar.dart';
 
 class HomeController extends GetxController {
   RxString title = 'Home'.obs;
   RxList<double> weights = [0.347, 0.653].obs;
-
-  RxList<Widget> children = <Widget>[
-    Sidebar(
-      child: SessionsList(),
-    ),
-    RightPane()
-  ].obs;
+  RxList<Widget> children = <Widget>[].obs;
 }
